+++ Source/Game/DataStorage/Structs/H_Records.cs
@@ -28,7 +28,7 @@ namespace Game.DataStorage
         public uint HolidayDescriptionID;
         public byte Priority;
         public sbyte CalendarFilterType;
-        public byte Flags;
+        public int Flags;
         public ushort[] Duration = new ushort[SharedConst.MaxHolidayDurations];
         public uint[] Date = new uint[SharedConst.MaxHolidayDates];                                 // dates in unix time starting at January, 1, 2000
         public byte[] CalendarFlags = new byte[SharedConst.MaxHolidayFlags];
